Restructure CMake project  - Moved header files to serparate include directory  - Added option to build tinyxml2 as external project at configure time  - Remove Find module for tinyxml2 as it already comes with an prefered CMake config file  - Remove extra target for shared and static libs as this can be controlled by option BUILD_SHARED_LIBS  - Export cmake config files as the prefered way to support CMake build systems  - Remove find module for this project as it is not needed anymore due to the exported config files  - Start to adapt CMake project to conform to modern cmake as described in Daniel Pfeifer's guideline slides 'CMAKE - Introduction and best practices (2015)' and 'Effective CMake - a random selection of best practices (2017)' which are summarized in https://gist.github.com/mbinna/c61dbb39bca0e4fb7d1f73b0d66a4fd1  - Add option for building the tests and copy the example files to the build directory. Thus the compiled test binary need not to be moved anymore.
