+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "tinyxml2"]
+	path = thirdparty/tinyxml2
+	url = https://github.com/leethomason/tinyxml2.git
