+++ README.md
@@ -1,6 +1,6 @@
 # tmxparser
 
-[![Build Status](https://travis-ci.org/sainteos/tmxparser.svg?branch=master)](https://travis-ci.org/sainteos/tmxparser)
+[![Build Status](https://travis-ci.org/tmxparser/tmxparser.svg?branch=master)](https://travis-ci.org/m-waka/tmxparser)
 
 This is a C++11 library for parsing the maps generated by
 [Tiled Map Editor](https://github.com/bjorn/tiled/).
@@ -25,7 +25,16 @@ An example file is provided to understand how to use the library.
 ## Dependencies
 
  * zlib
- * TinyXML2 >= 6.0.0
+ * TinyXML2 >= 6.0.0 
+
+## CMake options
+```
+USE_MINIZ         "Use miniz.c instead of zlib (default: OFF)"
+BUILD_SHARED_LIBS "Build shared libs, otherwise static libs. (default: ON)"
+BUILD_TINYXML2    "Build tinyxml2 as external project (default: OFF)"
+BUILD_TESTS       "Build tests. (default: OFF)"
+BUILD_DOCS        "Build documentation. (default: OFF)"
+```
 
 ## Installation
 
@@ -38,12 +47,11 @@ sudo make install
 ```
 
 ## To Run The Tests
-After make, there will be several artifacts within `build/`. One of which is called `run_tests`. The test runner expects to be in the `test/` folder so the links to reference files for the tests will be where it expects them to be.
+After make, there will be several artifacts within `build/`. One of which is called `run_tests`.
 ```
-...
+mkdir build
+cd build
+cmake .. -DBUILDTESTS=ON
 make
-sudo make install
-mv run_tests ../test/
-cd ../test/
 ./run_tests
 ```
