+++ cmake/tinyxml2.ext.in
@@ -1,0 +1,11 @@
+cmake_minimum_required(VERSION 3.5)
+
+project(ext_tinyxml2 NONE)
+
+include(ExternalProject)
+
+ExternalProject_Add(tinyxml2
+  SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/thirdparty/tinyxml2
+  INSTALL_DIR ${CMAKE_INSTALL_PREFIX}
+  CMAKE_ARGS -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR>
+  -DCMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE})
