+++ include/TmxEllipse.h
