+++ include/TmxImage.h
