+++ include/TmxImageLayer.h
