+++ include/TmxMapTile.h
