+++ include/TmxObject.h
