+++ include/TmxPolyline.h
