+++ include/TmxProperty.h
