+++ include/TmxPropertySet.h
