+++ include/TmxTerrainArray.h
