+++ include/TmxText.h
