+++ install_scripts/osx_install
@@ -1,4 +1,3 @@
 brew install zlib
 brew install doxygen
 brew install git-lfs
-brew install tinyxml2
