+++ src/base64/base64.cpp
@@ -25,7 +25,7 @@
 
 */
 
-#include "base64.h"
+#include <base64/base64.h>
 #include <iostream>
 #include <stdlib.h>
